requirements.txt
requests==2.32.3
gitpython